+++ Qb/Plugins/ColorTransform/include/colortransformelement.h
@@ -45,7 +45,7 @@ class ColorTransformElement: public QbElement
         Q_INVOKABLE QVariantList kernel() const;
 
     private:
-        QVector<float> m_kernel;
+        QVector<qreal> m_kernel;
         QbElementPtr m_convert;
 
     signals:
